Move isModifierCode to keycodes.js
